pyproject.toml
[project]
name = "pr-review-agent"
version = "0.1.0"
description = "Agentic PR review (Groq + Python CLI + VSCode extension)."
readme = "README.md"
authors = [{ name = "Unai Garay", email = "unaigaraymaestre@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "groq>=0.4.2",
    "pydantic>=2.6.1",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.1",
    "langgraph>=1.0.6",
    "langsmith>=0.6.3",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-community>=0.0.0",
    "duckduckgo-search>=5.0.0",
]

[project.scripts]
pr-agent = "pr_review_agent.cli:app"

[dependency-groups]
dev = ["pytest>=8.0.0", "ruff>=0.2.1", "mypy>=1.8.0", "types-requests>=2.31.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pr_review_agent"]
